Set-up live debug env
The minecraft_version variable was being used for grabbing alinLib
properly but because of that, selected the wrong minecraft version to
run.

I've made a new maven_minecraft_version variable for specifically the
purpose of grabbing maven dependencies because modrinth appears to use
the same syntax.

minecraft_version can now actually be used for the minecraft version

Also stopped using kels maven repo for grabbing files because I don't
need it